+++ src/main/java/net/geforcemods/securitycraft/datagen/RecipeGenerator.java
@@ -7,6 +7,7 @@ import net.geforcemods.securitycraft.SCTags;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.recipe.BlockReinforcingRecipe;
 import net.geforcemods.securitycraft.recipe.BlockUnreinforcingRecipe;
+import net.geforcemods.securitycraft.recipe.CopyPositionComponentItemRecipe;
 import net.geforcemods.securitycraft.recipe.LimitedUseKeycardRecipe;
 import net.geforcemods.securitycraft.util.Utils;
 import net.minecraft.core.HolderGetter;
@@ -52,6 +53,11 @@ public class RecipeGenerator extends RecipeProvider {
 		//Combine block with universal block reinforcer to unreinforce/reinforce it (depending on the reinforcer's mode), reducing the reinforcer's durability
 		SpecialRecipeBuilder.special(BlockReinforcingRecipe::new).save(output, scRecipeId("block_reinforcing"));
 		SpecialRecipeBuilder.special(BlockUnreinforcingRecipe::new).save(output, scRecipeId("block_unreinforcing"));
+		//tool copying
+		SpecialRecipeBuilder.special(CopyPositionComponentItemRecipe::cameraMonitor).save(output, scRecipeId("copy_camera_monitor"));
+		SpecialRecipeBuilder.special(CopyPositionComponentItemRecipe::mineRemoteAccessTool).save(output, scRecipeId("copy_mine_remote_access_tool"));
+		SpecialRecipeBuilder.special(CopyPositionComponentItemRecipe::sentryRemoteAccessTool).save(output, scRecipeId("copy_sentry_remote_access_tool"));
+		SpecialRecipeBuilder.special(CopyPositionComponentItemRecipe::sonicSecuritySystem).save(output, scRecipeId("copy_sonic_security_system"));
 		//combine keycard with limited use keycard to get keycards with a configurable limited amount of uses
 		SpecialRecipeBuilder.special(LimitedUseKeycardRecipe::new).save(output, scRecipeId("limited_use_keycards"));